"Chore: Update real estate app with latest transaction data
Refactor main.py to improve code maintainability and readability. Relocate logic, configurations, and classes into their respective modules under the 'src' directory. This makes it easier to update transaction data annually. The app now includes transactions up to the current year.

Note that it affects the Firebase, Azure, and Google BigQuery configurations, as they are part of the modules that have been relocated. Plus, these changes result in improved analytics tracking due to the included Firestore keys."
